+++ hardhat.config.ts
@@ -2,7 +2,8 @@ import '@nomicfoundation/hardhat-toolbox';
 import '@nomiclabs/hardhat-truffle5';
 import 'hardhat-dependency-compiler';
 import 'hardhat-deploy';
-import { HardhatUserConfig } from 'hardhat/config';
+import { TASK_COMPILE_SOLIDITY_GET_SOURCE_PATHS } from 'hardhat/builtin-tasks/task-names';
+import { HardhatUserConfig, subtask } from 'hardhat/config';
 import {
     HARDHAT_NETWORK_MNEMONIC,
     defaultHardhatNetworkParams,
@@ -192,5 +193,20 @@ const config: HardhatUserConfig = {
         keep: true, // Slither requires compiled dependencies
     },
 };
+
+/**
+ * Ignore some contracts from compilation. Credits to:
+ * https://kennysliding.medium.com/how-to-ignore-solidity-files-in-hardhat-compilation-6162963f8c84
+ */
+subtask(TASK_COMPILE_SOLIDITY_GET_SOURCE_PATHS).setAction(async (_, __, runSuper) => {
+    const paths = await runSuper();
+    // exclude paths that contain string
+    return await paths.filter(
+        (p: any) =>
+            !p.includes(
+                'IexecPoco1DelegateKYC', // Exclude KYC contracts until removal.
+            ),
+    );
+});
 
 export default config;